Adding basic support for reading an ARM64 memory dump (#1070)
* Adding basic support for reading an ARM64 memory dump

* Addressing initial CI build failures

* only parse config for aarch64 attributes if pagemode is aarch64

---------

+++ libvmi/private.h
@@ -105,9 +105,9 @@ struct vmi_instance {
         } x86;
 
         struct {
-            int t0sz;           /**< TTBR0 VA size (2^(64-t0sz)) */
+            uint64_t t0sz;           /**< TTBR0 VA size (2^(64-t0sz)) */
 
-            int t1sz;           /**< TTBR1 VA size (2^(64-t1sz)) */
+            uint64_t t1sz;           /**< TTBR1 VA size (2^(64-t1sz)) */
 
             page_size_t tg0;    /**< TTBR0 granule size: 4KB/16KB/64KB */
